refactor(serve): finish extracting the server implementation
