+++ scripts/serve-seam-allowlist.json
@@ -139,9 +139,8 @@
       "serve.RenderOutcome",
       "serve.ServeBuildHost",
       "serve.ServeBundleDaemon",
-      "serve.ServeDefaults",
+      "serve.ServeCommandOptions",
       "serve.ServeDiscovery",
-      "serve.ServeOptions",
       "serve.ServeRenderHost",
       "serve.ServeRunner",
       "serve.SvgOutcome"
@@ -155,7 +154,6 @@
       "serve.ServeAgentGrantStore",
       "serve.ServeBackgroundWork",
       "serve.ServeBundleHost",
-      "serve.ServeCatalogStore",
       "serve.ServeHttpServer",
       "serve.ServeImageStore",
       "serve.ServeImageUploadAuth",
@@ -163,8 +161,7 @@
       "serve.ServePreview",
       "serve.ServeRenderHost",
       "serve.ServeRunner",
-      "serve.ServeSessionRegistry",
-      "serve.ServeUrls"
+      "serve.ServeSessionRegistry"
     ]
   }
 }
